Revert "feat: add follower count for galleries (#12875)"
This reverts commit afc7ddf7af63b8379f749d8188a073b41243e09e.
