--- src/Utils/formatFollowerCount.ts
@@ -1,12 +1,0 @@
-export const formatFollowerCount = (n: number) => {
-  try {
-    const formatter = Intl.NumberFormat("en-US", {
-      notation: "compact",
-      compactDisplay: "short",
-    })
-
-    return formatter.format(n).toLocaleLowerCase()
-  } catch (error) {
-    return n
-  }
-}
